trigger deploy (#2)
- basic set up and content is displaying on page
- Added github folder and such
- added todo notes
- added todo notes round 2
- fixed tailwind rendering
- wraping up the sample blog

# Github Issue Number

#<IssueNumber>

# What is changing?

Describe what has changed as a part of this pull request.

# What does the change fix?

Describe the problem that is being solved as a part of this pull
request.

# Does this break existing functionality?

If this breaks existing functionality, provide details on what it breaks
and how you have accounted for the breakage.

# Does this require a version update?

Will there need to be a release, and if so, does it require a major
version change that needs to be communicated to consumers? Have
consumers been notified of the changes?

# How has this change been tested?

Describe the testing plan. This should include the following but isn't
limited to:

- [ ] Automated unit testing
- [ ] Automated integration testing
- [ ] Manual acceptance testing
- [ ] RRA Completed

# Additional details

Add any additional details that may be useful to reviewers here.

---

# Definition of Done

- [ ] Unit, acceptance, and integration tests added and passing
- [ ] Two sets of eyes on all code
- [ ] Design confirms design standards are met
- [ ] Product confirms acceptance criteria are met
- [ ] Deployed to highest pre-production environment
- [ ] When business metrics are impacted, the change is measurable and
existing measurements continue to work
- [ ] All guest-facing and employee-facing experiences are accessible
- [ ] The change has a work item, the commit references the work item,
and the change is recorded
- [ ] The collection, storage, use, and transmission of guest data
adheres to their explicit or implied consent and according to our
Privacy Notice
- [ ] Must comply with [OWASP](https://owasp.org/www-project- Please
note that each team can take this baseline definition and add to it, but
nothing can be removed from or altered on this list.